tools/epamedicationuniqueidentifier/requirements.txt
fhir.resources==7.1.0
lxml==5.2.2